Added user login and registration as well modified room.html

+++ base/views.py
@@ -1,11 +1,14 @@
 from django.shortcuts import render, redirect
 from django.db.models import Q
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
 from django.contrib import messages
 # Create your views here.
 from django.http import HttpResponse
 from .models import Room, Topic, User
 from .forms import RoomForm
+
 
 def loginPage(request):
     if request.user.is_authenticated:
@@ -13,7 +16,7 @@ def loginPage(request):
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
-        
+
         try:
             user = User.objects.get(username=username)
         except:
@@ -26,13 +29,35 @@ def loginPage(request):
         else:
             messages.error(request, 'Username or password is incorrect')
     context = {
-
+        'title': 'Login'
     }
     return render(request, 'base/login_register.html', context)
+
 
 def logoutUser(request):
     logout(request)
     return redirect('login')
+
+
+def registerPage(request):
+    context = {
+        'title': 'Register',
+        'form': UserCreationForm,
+    }
+    if request.user.is_authenticated:
+        return redirect('home')
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error has occurred during registration')
+    return render(request, 'base/login_register.html', context)
+
 
 def home(request):
     q = request.GET.get('q')
@@ -43,9 +68,7 @@ def home(request):
             Q(topics__description__icontains=q) |
             Q(name__icontains=q) |
             Q(description__icontains=q)
-        ).distinct(
-            # order by the number of topics that match the query
-        )
+        ).distinct()
     else:
         # get all the rooms
         rooms = Room.objects.all()
@@ -59,12 +82,15 @@ def home(request):
 
 def room(request, room_name):
     room = Room.objects.get(link=room_name)
+    messages = room.messages_set.all().order_by('-timestamp')
     context = {
         'room_name': room.description,
+        'messages': messages,
     }
     return render(request, 'base/room.html', context)
 
 
+@login_required(login_url='login')
 def room_form(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -80,9 +106,14 @@ def room_form(request):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not authorized to view this page')
+
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -94,8 +125,11 @@ def update_room(request, pk):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse('You are not authorized to view this page')
     if request.method == 'POST':
         room.delete()
         return redirect('home')
